I tried finding a webp-to-gif tool somewhere.  Anywhere.
No luck.  So then I tried to find scripts that I could link together. Not much luck there, but some.  Better than nothing.

So, this script is the first stage of a tool that I am now putting together for just that very narrow, specific purpose.  For when I have webp and need to convert them into gifs.  This script will tell me how many frames the webp file has, along with random bits of info about size, compression and dimensions.  Some combination of those should prove useful.

Another script, which has yet to be started, but should be simpler will turn the webp animation frames into individual webp images, which will be followed with yet another script to turn a webp image into some other form of image.  Probably just png, for now.

This collection of png images can then be bundled up into a gif with yet another script that will wrap gifski, which may already have a python library.

I believe gifski is open source, so that seems very possible.  But of course the libwebp library is also open source and so far, it is not so impressive on the external tools front.

I found a tool that should wrap libwebp for python and it is a wreched mess.  Python libraries should not ship with multiple copies of compiled binaries.  Which, for better or worse, this library does not.  Mine already looks more promising.

Can you believe that I started off today thinking about documentation for a node script and then messing around with GPS on Arduino.  I have not touched either of these and I stayed busy the bulk of the day.

Anyhow, to use this new python script set, you will need libwebp, from Google.  Looks like the source and compiled binaries have lived in a few places, but I found mine by starting here:
https://developers.google.com/speed/webp/docs/dwebp
Yes, it is the docs.  But it is a starting place.  Find the files to download.

The npm package animated-webp-info does some of the things that I was looking for and someone else said nice things about it in the forums.  My first script is an homage to this effort.  Although, mine is simpler, faster, and has fewer dependencies.  To compare for yourself look at the code at https://github.com/PeckZeg/animated-webp-info/blob/master/index.js.

Finally, webptools from scionoftech on github is the embarasing piece of absolute garbage that at first excited, then horrified, and eventually nauseated me into trying my own python scripts to do what I really though that so many others should have done by now.  Mine is barely started and already quite a bit more promising.  I should publish mine as a patch to his.  Best not to follow the link that comes next, just skip it and look at what I made instead.  Mine is better; how could it not be?

https://github.com/scionoftech/webptools/tree/master/lib/libwebp_win64